chore

+++ device/lib/audio_input/src/ADCSampler.cpp
@@ -1,3 +1,4 @@
+#include <Arduino.h>
 #include "ADCSampler.h"
 
 ADCSampler::ADCSampler(adc_unit_t adcUnit, adc1_channel_t adcChannel, const i2s_config_t &i2s_config) : I2SSampler(I2S_NUM_0, i2s_config)
@@ -20,15 +21,15 @@ void ADCSampler::unConfigureI2S()
     i2s_adc_disable(m_i2sPort);
 }
 
-int ADCSampler::read(int16_t *samples, int count)
+int ADCSampler::read(uint8_t *samples, int count)
 {
     // read from i2s
     size_t bytes_read = 0;
-    i2s_read(m_i2sPort, samples, sizeof(int16_t) * count, &bytes_read, portMAX_DELAY);
-    int samples_read = bytes_read / sizeof(int16_t);
+    i2s_read(m_i2sPort, samples, sizeof(uint8_t) * count, &bytes_read, portMAX_DELAY);
+    int samples_read = bytes_read / sizeof(uint8_t);
     for (int i = 0; i < samples_read; i++)
     {
-        samples[i] = (2048 - (uint16_t(samples[i]) & 0xfff)) * 15;
+        samples[i] = (2048 - (uint8_t(samples[i]) & 0xfff)) * 15;
     }
     return samples_read;
 }
