+++ device/lib/audio_input/src/ADCSampler.h
@@ -14,5 +14,5 @@ protected:
 
 public:
     ADCSampler(adc_unit_t adc_unit, adc1_channel_t adc_channel, const i2s_config_t &i2s_config);
-    virtual int read(int16_t *samples, int count);
+    virtual int read(uint8_t *samples, int count);
 };
